+++ vcfpy/__init__.py
@@ -13,15 +13,15 @@ from .header import Header, HeaderLine, SimpleHeaderLine, \
     PedigreeHeaderLine, SampleHeaderLine, CompoundHeaderLine, InfoHeaderLine, \
     FormatHeaderLine, SamplesInfos, FieldInfo, header_without_lines
 
-from .record import Record, Call, AltRecord, Substitution, SV, BreakEnd, \
+from .record import Record, Call, AltRecord, Substitution, BreakEnd, \
     SingleBreakEnd, SymbolicAllele
 
 from .record import SNV, MNV, DEL, INS, INDEL, SV, BND, SYMBOLIC, MIXED
-
 from .record import HOM_REF, HET, HOM_ALT
+from .record import FIVE_PRIME, THREE_PRIME, FORWARD, REVERSE
 
 from .reader import Reader
 
 from .writer import Writer
 
-__version__ = '0.5.0'
+__version__ = '0.6.0'