+++ HISTORY.rst
@@ -2,6 +2,14 @@
 History
 =======
 
+0.6.0 (2016-09-25
+-----------------
+
+* Refining implementation for breakend and symbolic allele class
+* Removing ``record.SV_CODES``
+* Refactoring parser module a bit to make the code cleaner
+* Fixing small typos and problems in documentation
+
 0.5.0 (2016-09-24)
 ------------------